Refactor newman_runner
