+++ crates/test_utils/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod connector_auth;
+pub mod newman_runner;
